+++ src/rcc.rs
@@ -429,6 +429,12 @@ impl Clocks {
     }
 }
 
+/// Common trait for most of peripherals
+pub trait Instance:
+    crate::Ptr + crate::Steal + Enable + Reset + RccBus<Bus: BusClock> + Deref<Target = Self::RB>
+{
+}
+
 /// Frequency on bus that peripheral is connected in
 pub trait BusClock {
     /// Calculates frequency depending on `Clock` state